[YS-416] refactor: 체크박스 컨테이너 컴포넌트 분리
